+++ carp_core/lib/carp_deployment/carp_core_deployment.dart
@@ -12,6 +12,8 @@
 /// definition in Kotlin.
 library carp_core_deployment;
 
+import 'dart:convert';
+
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
 import 'package:uuid/uuid.dart';