+++ carp_mobile_sensing/lib/runtime/runtime.dart
@@ -72,18 +72,18 @@ class SensingException implements Exception {
 
 /// A simple method for printing warning messages to the console.
 void info(String message) =>
-    (Settings().debugLevel.index >= DebugLevel.INFO.index)
+    (Settings().debugLevel.index >= DebugLevel.info.index)
         ? print('[CAMS INFO] $message')
         : 0;
 
 /// A simple method for printing warning messages to the console.
 void warning(String message) =>
-    (Settings().debugLevel.index >= DebugLevel.WARNING.index)
+    (Settings().debugLevel.index >= DebugLevel.warning.index)
         ? print('[CAMS WARNING]  $message')
         : 0;
 
 /// A simple method for printing debug messages to the console.
 void debug(String message) =>
-    (Settings().debugLevel.index >= DebugLevel.DEBUG.index)
+    (Settings().debugLevel.index >= DebugLevel.debug.index)
         ? print('[CAMS DEBUG] $message')
         : 0;